requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
tmdbv3api==1.9.0
google-generativeai==0.3.2
gunicorn==21.2.0
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.0.3
pytz==2023.3 